renamed items functions to inventory functions
