+++ src/components/CraftButton.vue
@@ -3,7 +3,7 @@ import { usePlannerStore } from '../store/planner-store';
 import { Item } from '../types/outputdata';
 import { computed } from 'vue';
 import { storeToRefs } from 'pinia';
-import { canCraft } from '../store/store-item-functions';
+import { canCraft } from '../store/store-inventory-functions';
 
 const { craftItem } = usePlannerStore();
 const { inventory } = storeToRefs(usePlannerStore());
