+++ src/store/store-inventory-functions.ts
